Yesod.RssFeed

Documentation

rssFeed :: MonadHandler m => Feed (Route (HandlerSite m)) -> m RepRss

Generate the feed

rssFeedText :: MonadHandler m => Feed Text -> m RepRss

Same as rssFeed but for 'Feed Text'. Useful for cases where you are generating a feed of external links.

rssLink

Arguments

:: MonadWidget m

=> Route (HandlerSite m)

-> Text
title

-> m ()

Generates a link tag in the head of a widget.

newtype RepRss

Constructors

RepRss Content

data Feed url

The overall feed

Constructors

Feed

Fields

feedTitle :: Text

feedLinkSelf :: url

feedLinkHome :: url

feedAuthor :: Text

feedDescription :: Html

note: currently only used for Rss

feedLanguage :: Text

note: currently only used for Rss, possible values: http://www.rssboard.org/rss-language-codes

feedUpdated :: UTCTime

feedLogo :: Maybe (url, Text)

feedEntries :: [FeedEntry url]

data FeedEntry url

Each feed entry

Constructors

FeedEntry

Fields

feedEntryLink :: url

feedEntryUpdated :: UTCTime

feedEntryTitle :: Text

feedEntryContent :: Html

feedEntryEnclosure :: Maybe (EntryEnclosure url)

Allows enclosed data: RSS <enclosure> or Atom <link rel=enclosure>

Since: 1.5

data EntryEnclosure url

RSS and Atom allow for linked content to be enclosed in a feed entry. This represents the enclosed content.

Atom feeds ignore enclosedSize and enclosedMimeType.

Since: 1.6

Constructors

EntryEnclosure

Fields

enclosedUrl :: url

enclosedSize :: Int

Specified in bytes

enclosedMimeType :: Text